requirements.txt
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.3.0
scipy>=1.7.0
openai>=0.27.0
scikit-learn>=0.24.0
matlabengine>=9.12.0
python-dotenv>=0.19.0 